Se añaden los metodos de la entidad Pais

+++ src/main/java/com/flyease/flyeaseapirest/controller/PaisController.java
@@ -4,6 +4,7 @@ import com.flyease.flyeaseapirest.exception.BadRequestException;
 import com.flyease.flyeaseapirest.exception.ResourceNotFoundException;
 import com.flyease.flyeaseapirest.model.dto.PaisDto;
 import com.flyease.flyeaseapirest.model.entity.Pais;
+import com.flyease.flyeaseapirest.model.payload.ApiResponse;
 import com.flyease.flyeaseapirest.model.payload.MensajeResponse;
 import com.flyease.flyeaseapirest.service.IPaisService;
 import jakarta.validation.Valid;
@@ -16,7 +17,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/v2")
+@RequestMapping("/ApiFlyEase/v2")
 public class PaisController {
 
     @Autowired
@@ -31,21 +32,43 @@ public class PaisController {
 
         return new ResponseEntity<>(
                 MensajeResponse.builder()
-                        .mensaje("")
-                        .object(getList)
+                        .mensaje("ok")
+                        .sucess(true)
+                        .response(getList)
                         .build()
                 , HttpStatus.OK);
     }
 
+    @GetMapping("paises/{id}")
+    public ResponseEntity<?> showById(@PathVariable Integer id) {
+        Pais pais = paisService.findById(id);
 
-    @PostMapping("pais")
+        if (pais == null) {
+            throw new ResourceNotFoundException("pais", "id", id);
+        }
+
+        return new ResponseEntity<>(
+                MensajeResponse.builder()
+                        .mensaje("ok")
+                        .sucess(true)
+                        .response(PaisDto.builder()
+                                .idPais(pais.getIdPais())
+                                .nombre(pais.getNombre())
+                                .fechaRegistro(pais.getFechaRegistro())
+                                .build())
+                        .build()
+                , HttpStatus.OK);
+    }
+
+    @PostMapping("paises")
     public ResponseEntity<?> create(@RequestBody @Valid PaisDto paisDto) {
         Pais paisSave = null;
         try {
             paisSave = paisService.save(paisDto);
             return new ResponseEntity<>(MensajeResponse.builder()
                     .mensaje("Guardado correctamente")
-                    .object(PaisDto.builder()
+                    .sucess(true)
+                    .response(PaisDto.builder()
                             .idPais(paisSave.getIdPais())
                             .nombre(paisSave.getNombre() )
                             .fechaRegistro(paisSave.getFechaRegistro())
@@ -57,16 +80,17 @@ public class PaisController {
         }
     }
 
-    @PutMapping("pais/{id}")
+    @PutMapping("paises/{id}")
     public ResponseEntity<?> update(@RequestBody @Valid PaisDto paisDto, @PathVariable Integer id) {
         Pais paisUpdate = null;
         try {
             if (paisService.existsById(id)) {
                 paisDto.setIdPais(id);
-                paisUpdate = paisService.save(paisDto);
+                paisUpdate = paisService.update(paisDto);
                 return new ResponseEntity<>(MensajeResponse.builder()
-                        .mensaje("Guardado correctamente")
-                        .object(PaisDto.builder()
+                        .mensaje("Actualizado correctamente")
+                        .sucess(true)
+                        .response(PaisDto.builder()
                                 .idPais(paisUpdate.getIdPais())
                                 .nombre(paisUpdate.getNombre())
                                 .fechaRegistro(paisUpdate.getFechaRegistro())
@@ -81,35 +105,24 @@ public class PaisController {
         }
     }
 
-    @DeleteMapping("pais/{id}")
+    @DeleteMapping("paises/{id}")
     public ResponseEntity<?> delete(@PathVariable Integer id) {
         try {
             Pais paisDelete = paisService.findById(id);
-            paisService.delete(paisDelete);
-            return new ResponseEntity<>(paisDelete, HttpStatus.NO_CONTENT);
+            if (paisDelete != null) {
+                paisService.delete(paisDelete);
+                return new ResponseEntity<>(ApiResponse.builder()
+                        .mensaje("Eliminado correctamente")
+                        .sucess(true)
+                        .build()
+                        , HttpStatus.OK);
+            }
+            else {
+                throw new ResourceNotFoundException("pais", "id", id);
+            }
         } catch (DataAccessException exDt) {
             throw  new BadRequestException(exDt.getMessage());
         }
-    }
-
-    @GetMapping("pais/{id}")
-    public ResponseEntity<?> showById(@PathVariable Integer id) {
-        Pais pais = paisService.findById(id);
-
-        if (pais == null) {
-            throw new ResourceNotFoundException("pais", "id", id);
-        }
-
-        return new ResponseEntity<>(
-                MensajeResponse.builder()
-                        .mensaje("")
-                        .object(PaisDto.builder()
-                                .idPais(pais.getIdPais())
-                                .nombre(pais.getNombre())
-                                .fechaRegistro(pais.getFechaRegistro())
-                                .build())
-                        .build()
-                , HttpStatus.OK);
     }
 
 }
